requirements.txt
# Core FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# Firebase Admin
firebase-admin>=6.2.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Data Validation and Serialization
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Async and HTTP
httpx>=0.27.0
websockets>=12.0
python-dateutil>=2.8.2

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
orjson>=3.9.0
typing-extensions>=4.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Database (if needed)
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9  # For PostgreSQL
